Moved isSpringBootSanityTest method
